Module rustybot_core::directive

This module contains the base trait for all Directive and all necessary conveniences to allow for user-configuration of directive defaults.

Structs

DirectiveData

A struct that contains the default settings for a Directive and the name it takes in configuration sources. The name must be unique.

Enums

Setting

Represents a value for a setting

Traits

Directive

A parser for action steps, each directive represents a type of Action.

Functions

initialize_settings_object

Returns a Settings object from an array as a bit of syntactic sugar

Type Definitions

Settings

The Settings object is a hashmap of option names to a default value